+++ mcp_server/tests/test_session_isolation.py
@@ -19,10 +19,10 @@ import sys
 PROJECT_ROOT = Path(__file__).parent.parent.parent
 sys.path.insert(0, str(PROJECT_ROOT))
 
-from mcp_server.server.main import (
-    _get_session, _set_client_roots, execute_python_code, 
-    get_presentation_tree, _session_store, SessionContext
+from mcp_server.server.session import (
+    get_session, set_client_roots, get_session_store, clear_session_store, SessionContext
 )
+from mcp_server.server.tools import execute_python_code, get_presentation_tree
 from mcp import types
 
 
@@ -31,14 +31,14 @@ class TestSessionIsolation:
     
     def setup_method(self):
         """Clear session store before each test."""
-        _session_store.clear()
+        clear_session_store()
     
     @pytest.mark.asyncio
     async def test_concurrent_sessions_isolation(self):
         """Test that concurrent sessions maintain separate state."""
         
         # Mock pptx library
-        with patch('mcp_server.server.main.pptx', Mock()):
+        with patch('mcp_server.server.session.pptx', Mock()):
             results = []
             errors = []
             
@@ -46,8 +46,8 @@ class TestSessionIsolation:
                 """Worker function to simulate a client session."""
                 try:
                     # Force this thread to use a specific session ID
-                    import mcp_server.server.main as main_module
-                    main_module._current_session_id.value = session_id
+                    import mcp_server.server.session as session_module
+                    session_module._current_session_id.value = session_id
                     
                     # Create session context
                     session = SessionContext(
@@ -65,10 +65,11 @@ class TestSessionIsolation:
                         session.loaded_presentation_path = Path(f"/test/{session_id}.pptx")
                     
                     # Store session
-                    _session_store[session_id] = session
+                    session_store = get_session_store()
+                    session_store[session_id] = session
                     
                     # Get session and verify it's the right one
-                    retrieved_session = _get_session()
+                    retrieved_session = get_session()
                     results.append({
                         'session_id': session_id,
                         'retrieved_session_id': retrieved_session.session_id,
@@ -145,26 +146,27 @@ class TestSessionIsolation:
             client_roots=[]
         )
         
-        _session_store["old_session"] = old_session
-        _session_store["recent_session"] = recent_session
+        session_store = get_session_store()
+        session_store["old_session"] = old_session
+        session_store["recent_session"] = recent_session
         
         # Verify both sessions exist
-        assert len(_session_store) == 2
+        assert len(session_store) == 2
         
         # Import and call cleanup function with 1 hour max age
-        from mcp_server.server.main import _cleanup_expired_sessions
-        _cleanup_expired_sessions(max_age=3600)  # 1 hour
+        from mcp_server.server.session import cleanup_expired_sessions
+        cleanup_expired_sessions(max_age=3600)  # 1 hour
         
         # Verify old session was removed but recent session remains
-        assert "old_session" not in _session_store
-        assert "recent_session" in _session_store
-        assert len(_session_store) == 1
+        assert "old_session" not in session_store
+        assert "recent_session" in session_store
+        assert len(session_store) == 1
     
     @pytest.mark.asyncio
     async def test_tools_use_correct_session(self):
         """Test that tools access the correct session's data."""
         
-        with patch('mcp_server.server.main.pptx', Mock()):
+        with patch('mcp_server.server.session.pptx', Mock()):
             # Create two sessions with different states
             session_1 = SessionContext(
                 session_id="session_1",
@@ -186,11 +188,12 @@ class TestSessionIsolation:
             session_1.loaded_presentation = mock_presentation
             session_1.loaded_presentation_path = Path("/test/session1.pptx")
             
-            _session_store["session_1"] = session_1
-            _session_store["session_2"] = session_2
+            session_store = get_session_store()
+            session_store["session_1"] = session_1
+            session_store["session_2"] = session_2
             
             # Test session 1 can execute code
-            with patch('mcp_server.server.main._get_session', return_value=session_1):
+            with patch('mcp_server.server.tools.get_session', return_value=session_1):
                 result = await execute_python_code("print('Hello from session 1')")
                 import json
                 result_data = json.loads(result)
@@ -198,7 +201,7 @@ class TestSessionIsolation:
                 assert "Hello from session 1" in result_data["stdout"]
             
             # Test session 2 cannot execute code (no presentation)
-            with patch('mcp_server.server.main._get_session', return_value=session_2):
+            with patch('mcp_server.server.tools.get_session', return_value=session_2):
                 result = await execute_python_code("print('Hello from session 2')")
                 result_data = json.loads(result)
                 assert result_data["success"] is False
